[OAuth] Request all supported scopes when not using the OAuth debugger
When using the OAuth debugger, it correctly adds the scope search
parameter with the ones provided in the `scopes_supported` key.

However, the main "connect" button does not pass them to the MCP SDK and
they are missing from the call. Some OAuth providers and setups will
require you to have them and the OAuth flow fails in these cases. This
change fixes that.